requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiolimiter==1.2.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Authlib==1.3.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
fastapi==0.115.12
filelock==3.18.0
frozenlist==1.5.0
fsspec==2025.3.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.2
grpcio==1.71.0
grpcio-health-checking==1.71.0
grpcio-status==1.71.0
grpcio-tools==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
langdetect==1.0.9
lxml==5.3.1
multidict==6.4.3
numpy==2.2.4
packaging==24.2
pillow==11.2.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.2
python-docx==1.1.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
setuptools==77.0.3
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tenacity==9.1.2
tokenizers==0.21.1
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.2
validators==0.34.0
voyageai==0.3.2
weaviate-client==4.11.2
yarl==1.19.0
